requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.3.4
psycopg[binary,pool]==3.2.1
numpy==1.26.4
Pillow==10.3.0
onnxruntime==1.18.1
httpx==0.27.0
